requirements.txt
absl-py==0.10.0
aiohttp==3.6.3
argcomplete==1.12.3
astetik==1.13
astor==0.8.1
async-timeout==3.0.1
attrs==20.2.0
backcall==0.2.0
blinker==1.4
brotlipy==0.7.0
cachetools==4.1.1
certifi==2021.5.30
cffi==1.14.3
chances==0.1.9
chardet==3.0.4
click==7.1.2
cloudpickle==2.1.0
colorama==0.4.4
cryptography==3.1.1
cycler==0.10.0
debugpy==1.4.1
decorator==5.0.9
entrypoints==0.3
fastjsonschema==2.16.1
future==0.18.2
gast==0.2.2
geonamescache==1.4.0
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
graphviz==0.17
grpcio==1.31.0
h5py==2.10.0
hyperopt==0.2.5
idna==2.10
importlib-metadata==2.0.0
importlib-resources==5.9.0
ipykernel==6.2.0
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jsonschema==4.7.2
jupyter-client==7.0.1
jupyter-core==4.7.1
Keras==2.3.1
Keras-Applications==1.0.8
keras-hypetune==0.2.1
Keras-Preprocessing==1.1.0
kerasplotlib==1.0
kiwisolver==1.3.1
Markdown==3.3.2
matplotlib==3.4.3
matplotlib-inline==0.1.2
mkl-fft==1.2.0
mkl-random==1.0.4
mkl-service==2.3.0
multidict==4.7.6
nbformat==5.4.0
nest-asyncio==1.5.1
networkx==2.6.3
numpy==1.19.1
oauthlib==3.1.0
opencv-python==4.5.3.56
opt-einsum==3.1.0
packaging==21.3
pandas==1.2.4
parso==0.8.2
patsy==0.5.2
pickleshare==0.7.5
Pillow==8.3.1
pip==20.2.4
plotly==5.3.1
prompt-toolkit==3.0.20
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pydotplus==2.0.2
Pygments==2.10.0
PyJWT==1.7.1
pynput==1.7.6
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2021.1
pywin32==301
PyYAML==5.3.1
pyzmq==22.2.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.24.2
scipy==1.4.1
seaborn==0.11.2
setuptools==50.3.0.post20201006
six==1.15.0
sklearn==0.0
statsmodels==0.13.2
talos==1.3
tenacity==8.0.1
tensorboard==2.1.1
tensorboard-plugin-wit==1.6.0
tensorflow==2.1.0
tensorflow-estimator==2.1.0
termcolor==1.1.0
threadpoolctl==2.2.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
typing-extensions==4.3.0
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==0.14.1
wheel==0.35.1
win-inet-pton==1.1.0
wincertstore==0.2
wrangle==0.7.2
wrapt==1.12.1
yarl==1.6.2
zipp==3.3.1
mediapipe
mss
pyautogui
pytesseract
pygetwindow
Flask